// apps/crm/client/src/components/tanstack-table/cards/contacts/TableContactsOptionBtn.module.scss.d.ts
export type Styles = {
  contactsOptionBtn: string;
  contactsOptionBtn__menu: string;
  contactsOptionBtn__menuItem: string;
  contactsOptionBtn__menuItem__svg: string;
  contactsOptionBtn__menuItemWarning: string;
  contactsOptionBtn__menuItemWarning__svg: string;
  contactsOptionBtn__popover: string;
  contactsOptionBtn__svg: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;